CI: Pin the Rust version for clippy
We use the latest stable Rust in CI for rustfmt, clippy, and tests. This
has the downside that PRs are rejected when there is a new clippy
version with new lints that fire in unrelated parts of the project.

This change pins the clippy job to a particular Rust version. We
can then update clippy (and fix issues) separately from other PRs.